[Refactor] 1차 리팩토링: 로직 분리, 컴포넌트 세분화, 스타일 속성 함수 생성
