+++ app/page.tsx
@@ -1,6 +1,6 @@
 export const metadata = {
   title: '카페 마스터즈',
-  description: '카페를 수집하고 다니세요!',
+  description: '카페를 수집하여 카드를 모아 카페 마스터가 되어보세요',
   keywords: ['카페 마스터즈', '카페', '카페 추천', '서울 카페'],
   openGraph: {
     // 내용
