+++ components/search.tsx
@@ -5,6 +5,7 @@ import { useRouter } from 'next/navigation';
 import { useMapStore, useCheckStore } from 'utils/store';
 
 export default function Search() {
+  const router = useRouter();
   const keyword = useMapStore((state) => state.keyword);
   const setKeyword = useMapStore((state) => state.setKeyword);
   const setIsSubSidebarOpen = useCheckStore(
@@ -13,8 +14,6 @@ export default function Search() {
   const isDarkTheme = useCheckStore((state) => state.isDarkTheme);
 
   const [localKeyword, setLocalKeyword] = useState(keyword);
-
-  const router = useRouter();
 
   const handleSearch = () => {
     setKeyword(localKeyword);
@@ -38,7 +37,7 @@ export default function Search() {
         <div className="relative">
           <input
             className={`w-full bg-transparent ${isDarkTheme ? 'placeholder:text-gray-200 text-white' : 'placeholder:text-slate-400 text-slate-700'} text-md border border-slate-200 rounded-md pl-3 pr-28 py-4 transition duration-300 ease focus:outline-none focus:border-main hover:border-slate-300 shadow-sm focus:shadow`}
-            placeholder="찾으시는 곳"
+            placeholder="ex) 성수, 동성로"
             value={localKeyword}
             onChange={(e) => setLocalKeyword(e.target.value)}
             onKeyDown={handleKeyDown}
